refactor: move utils up, adjust comments
