+++ example/src/Examples/AnimatedFABExample/AnimatedFABExample.tsx
@@ -101,7 +101,7 @@ const AnimatedFABExample = () => {
         </View>
       );
     },
-    [visible]
+    [visible, isV3]
   );
 
   const onScroll = ({
@@ -142,14 +142,7 @@ const AnimatedFABExample = () => {
               : colors?.background,
           },
         ]}
-        contentContainerStyle={[
-          styles.container,
-          {
-            backgroundColor: isV3
-              ? (md('md.sys.color.background') as string)
-              : colors?.background,
-          },
-        ]}
+        contentContainerStyle={styles.container}
         onScroll={onScroll}
       />
