--- src/components/FAB/AnimatedFAB/index.ts
@@ -1,3 +1,0 @@
-export { default } from './AnimatedFAB';
-export * from './AnimatedFAB';
-export * from './utils';
